+++ packages/core/src/__tests__/session-manager/spawn.test.ts
@@ -1,0 +1,979 @@
+import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
+import {
+  mkdirSync,
+  readFileSync,
+} from "node:fs";
+import { join } from "node:path";
+import { createSessionManager } from "../../session-manager.js";
+import {
+  writeMetadata,
+  readMetadata,
+  readMetadataRaw,
+  updateMetadata,
+} from "../../metadata.js";
+import type {
+  OrchestratorConfig,
+  PluginRegistry,
+  Runtime,
+  Agent,
+  Workspace,
+  Tracker,
+} from "../../types.js";
+import { setupTestContext, teardownTestContext, makeHandle, type TestContext } from "./test-utils.js";
+import { installMockOpencode, installMockGit } from "./opencode-helpers.js";
+
+let ctx: TestContext;
+let tmpDir: string;
+let sessionsDir: string;
+let mockRuntime: Runtime;
+let mockAgent: Agent;
+let mockWorkspace: Workspace;
+let mockRegistry: PluginRegistry;
+let config: OrchestratorConfig;
+let originalPath: string | undefined;
+
+beforeEach(() => {
+  ctx = setupTestContext();
+  ({ tmpDir, sessionsDir, mockRuntime, mockAgent, mockWorkspace, mockRegistry, config, originalPath } = ctx);
+});
+
+afterEach(() => {
+  teardownTestContext(ctx);
+});
+
+describe("spawn", () => {
+  it("creates a session with workspace, runtime, and agent", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app" });
+
+    expect(session.id).toBe("app-1");
+    expect(session.status).toBe("spawning");
+    expect(session.projectId).toBe("my-app");
+    expect(session.runtimeHandle).toEqual(makeHandle("rt-1"));
+
+    // Verify workspace was created
+    expect(mockWorkspace.create).toHaveBeenCalled();
+    // Verify agent launch command was requested
+    expect(mockAgent.getLaunchCommand).toHaveBeenCalled();
+    // Verify runtime was created
+    expect(mockRuntime.create).toHaveBeenCalled();
+  });
+
+  it("blocks spawn while the project is globally paused", async () => {
+    writeMetadata(sessionsDir, "app-orchestrator", {
+      worktree: join(tmpDir, "my-app"),
+      branch: "main",
+      status: "working",
+      role: "orchestrator",
+      project: "my-app",
+      runtimeHandle: JSON.stringify(makeHandle("rt-orchestrator")),
+    });
+    updateMetadata(sessionsDir, "app-orchestrator", {
+      globalPauseUntil: new Date(Date.now() + 60_000).toISOString(),
+      globalPauseReason: "Rate limit reached",
+      globalPauseSource: "app-9",
+    });
+
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    await expect(sm.spawn({ projectId: "my-app" })).rejects.toThrow(
+      "Project is paused due to model rate limit until",
+    );
+    expect(mockRuntime.create).not.toHaveBeenCalled();
+  });
+
+  it("uses issue ID to derive branch name", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-100" });
+
+    expect(session.branch).toBe("feat/INT-100");
+    expect(session.issueId).toBe("INT-100");
+  });
+
+  it("sanitizes free-text issueId into a valid branch slug", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "fix login bug" });
+
+    expect(session.branch).toBe("feat/fix-login-bug");
+  });
+
+  it("preserves casing for branch-safe issue IDs without tracker", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-9999" });
+
+    expect(session.branch).toBe("feat/INT-9999");
+  });
+
+  it("sanitizes issueId with special characters", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({
+      projectId: "my-app",
+      issueId: "Fix: user can't login (SSO)",
+    });
+
+    expect(session.branch).toBe("feat/fix-user-can-t-login-sso");
+  });
+
+  it("truncates long slugs to 60 characters", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({
+      projectId: "my-app",
+      issueId:
+        "this is a very long issue description that should be truncated to sixty characters maximum",
+    });
+
+    expect(session.branch!.replace("feat/", "").length).toBeLessThanOrEqual(60);
+  });
+
+  it("does not leave trailing dash after truncation", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    // Craft input where the 60th char falls on a word boundary (dash)
+    const session = await sm.spawn({
+      projectId: "my-app",
+      issueId: "ab ".repeat(30), // "ab ab ab ..." → "ab-ab-ab-..." truncated at 60
+    });
+
+    const slug = session.branch!.replace("feat/", "");
+    expect(slug).not.toMatch(/-$/);
+    expect(slug).not.toMatch(/^-/);
+  });
+
+  it("falls back to sessionId when issueId sanitizes to empty string", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "!!!" });
+
+    // Slug is empty after sanitization, falls back to sessionId
+    expect(session.branch).toMatch(/^feat\/app-\d+$/);
+  });
+
+  it("sanitizes issueId containing '..' (invalid in git branch names)", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "foo..bar" });
+
+    // '..' is invalid in git refs, so it should be slugified
+    expect(session.branch).toBe("feat/foo-bar");
+  });
+
+  it("uses tracker.branchName when tracker is available", async () => {
+    const mockTracker: Tracker = {
+      name: "mock-tracker",
+      getIssue: vi.fn().mockResolvedValue({}),
+      isCompleted: vi.fn().mockResolvedValue(false),
+      issueUrl: vi.fn().mockReturnValue(""),
+      branchName: vi.fn().mockReturnValue("custom/INT-100-my-feature"),
+      generatePrompt: vi.fn().mockResolvedValue(""),
+    };
+
+    const registryWithTracker: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return mockAgent;
+        if (slot === "workspace") return mockWorkspace;
+        if (slot === "tracker") return mockTracker;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({
+      config,
+      registry: registryWithTracker,
+    });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-100" });
+    expect(session.branch).toBe("custom/INT-100-my-feature");
+  });
+
+  it("increments session numbers correctly", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    // Pre-create some metadata to simulate existing sessions
+    writeMetadata(sessionsDir, "app-3", { worktree: "/tmp", branch: "b", status: "working" });
+    writeMetadata(sessionsDir, "app-7", { worktree: "/tmp", branch: "b", status: "working" });
+
+    const session = await sm.spawn({ projectId: "my-app" });
+    expect(session.id).toBe("app-8");
+  });
+
+  it("does not reuse a killed session branch on recreate", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const first = await sm.spawn({ projectId: "my-app" });
+    expect(first.id).toBe("app-1");
+    expect(first.branch).toBe("session/app-1");
+
+    await sm.kill(first.id);
+
+    const second = await sm.spawn({ projectId: "my-app" });
+    expect(second.id).toBe("app-2");
+    expect(second.branch).toBe("session/app-2");
+  });
+
+  it("skips remote session branches when allocating a fresh session id", async () => {
+    const mockGitBin = installMockGit(tmpDir, ["session/app-22"]);
+    process.env.PATH = `${mockGitBin}:${originalPath ?? ""}`;
+    mkdirSync(config.projects["my-app"]!.path, { recursive: true });
+
+    const sm = createSessionManager({ config, registry: mockRegistry });
+    const session = await sm.spawn({ projectId: "my-app" });
+
+    expect(session.id).toBe("app-23");
+    expect(session.branch).toBe("session/app-23");
+  });
+
+  it("writes metadata file", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+    await sm.spawn({ projectId: "my-app", issueId: "INT-42" });
+
+    const meta = readMetadata(sessionsDir, "app-1");
+    expect(meta).not.toBeNull();
+    expect(meta!.status).toBe("spawning");
+    expect(meta!.project).toBe("my-app");
+    expect(meta!.issue).toBe("INT-42");
+  });
+
+  it("reuses OpenCode session mapping by issue when available", async () => {
+    const opencodeAgent: Agent = {
+      ...mockAgent,
+      name: "opencode",
+    };
+
+    const registryWithOpenCode: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return opencodeAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    config = {
+      ...config,
+      defaults: { ...config.defaults, agent: "opencode" },
+      projects: {
+        ...config.projects,
+        "my-app": {
+          ...config.projects["my-app"],
+          agent: "opencode",
+        },
+      },
+    };
+
+    writeMetadata(sessionsDir, "app-9", {
+      worktree: "/tmp/old",
+      branch: "feat/INT-42",
+      status: "killed",
+      project: "my-app",
+      issue: "INT-42",
+      agent: "opencode",
+      createdAt: "2026-01-01T00:00:00.000Z",
+      opencodeSessionId: "ses_existing",
+    });
+
+    const sm = createSessionManager({ config, registry: registryWithOpenCode });
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-42" });
+
+    expect(opencodeAgent.getLaunchCommand).toHaveBeenCalledWith(
+      expect.objectContaining({
+        projectConfig: expect.objectContaining({
+          agentConfig: expect.objectContaining({ opencodeSessionId: "ses_existing" }),
+        }),
+      }),
+    );
+
+    const metadata = readMetadataRaw(sessionsDir, session.id);
+    expect(metadata?.["opencodeSessionId"]).toBe("ses_existing");
+  });
+
+  it("reuses most recent session-id candidate without relying on timestamps", async () => {
+    const opencodeAgent: Agent = {
+      ...mockAgent,
+      name: "opencode",
+    };
+
+    const registryWithOpenCode: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return opencodeAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    config = {
+      ...config,
+      defaults: { ...config.defaults, agent: "opencode" },
+      projects: {
+        ...config.projects,
+        "my-app": {
+          ...config.projects["my-app"],
+          agent: "opencode",
+        },
+      },
+    };
+
+    writeMetadata(sessionsDir, "app-1", {
+      worktree: "/tmp/old-no-ts",
+      branch: "feat/INT-42",
+      status: "killed",
+      project: "my-app",
+      issue: "INT-42",
+      agent: "opencode",
+      opencodeSessionId: "ses_invalid_ts",
+    });
+
+    writeMetadata(sessionsDir, "app-2", {
+      worktree: "/tmp/new-with-ts",
+      branch: "feat/INT-42",
+      status: "killed",
+      project: "my-app",
+      issue: "INT-42",
+      agent: "opencode",
+      opencodeSessionId: "ses_valid_newer",
+    });
+
+    const sm = createSessionManager({ config, registry: registryWithOpenCode });
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-42" });
+
+    expect(opencodeAgent.getLaunchCommand).toHaveBeenCalledWith(
+      expect.objectContaining({
+        projectConfig: expect.objectContaining({
+          agentConfig: expect.objectContaining({ opencodeSessionId: "ses_valid_newer" }),
+        }),
+      }),
+    );
+
+    const metadata = readMetadataRaw(sessionsDir, session.id);
+    expect(metadata?.["opencodeSessionId"]).toBe("ses_valid_newer");
+  });
+
+  it("does not reuse issue mapping when opencodeIssueSessionStrategy is ignore", async () => {
+    const opencodeAgent: Agent = {
+      ...mockAgent,
+      name: "opencode",
+    };
+
+    const registryWithOpenCode: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return opencodeAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    config = {
+      ...config,
+      defaults: { ...config.defaults, agent: "opencode" },
+      projects: {
+        ...config.projects,
+        "my-app": {
+          ...config.projects["my-app"],
+          agent: "opencode",
+          opencodeIssueSessionStrategy: "ignore",
+        },
+      },
+    };
+
+    writeMetadata(sessionsDir, "app-9", {
+      worktree: "/tmp/old",
+      branch: "feat/INT-42",
+      status: "killed",
+      project: "my-app",
+      issue: "INT-42",
+      agent: "opencode",
+      opencodeSessionId: "ses_existing",
+    });
+
+    const sm = createSessionManager({ config, registry: registryWithOpenCode });
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-42" });
+
+    expect(opencodeAgent.getLaunchCommand).toHaveBeenCalledWith(
+      expect.objectContaining({
+        projectConfig: expect.objectContaining({
+          agentConfig: expect.not.objectContaining({ opencodeSessionId: expect.any(String) }),
+        }),
+      }),
+    );
+
+    const metadata = readMetadataRaw(sessionsDir, session.id);
+    expect(metadata?.["opencodeSessionId"]).toBeUndefined();
+  });
+
+  it("deletes old issue mappings and starts fresh when opencodeIssueSessionStrategy is delete", async () => {
+    const deleteLogPath = join(tmpDir, "opencode-delete-issue.log");
+    const mockBin = installMockOpencode(tmpDir, "[]", deleteLogPath);
+    process.env.PATH = `${mockBin}:${originalPath ?? ""}`;
+
+    const opencodeAgent: Agent = {
+      ...mockAgent,
+      name: "opencode",
+    };
+
+    const registryWithOpenCode: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return opencodeAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    config = {
+      ...config,
+      defaults: { ...config.defaults, agent: "opencode" },
+      projects: {
+        ...config.projects,
+        "my-app": {
+          ...config.projects["my-app"],
+          agent: "opencode",
+          opencodeIssueSessionStrategy: "delete",
+        },
+      },
+    };
+
+    writeMetadata(sessionsDir, "app-8", {
+      worktree: "/tmp/old1",
+      branch: "feat/INT-42",
+      status: "killed",
+      project: "my-app",
+      issue: "INT-42",
+      agent: "opencode",
+      opencodeSessionId: "ses_old_1",
+    });
+    writeMetadata(sessionsDir, "app-9", {
+      worktree: "/tmp/old2",
+      branch: "feat/INT-42",
+      status: "killed",
+      project: "my-app",
+      issue: "INT-42",
+      agent: "opencode",
+      opencodeSessionId: "ses_old_2",
+    });
+
+    const sm = createSessionManager({ config, registry: registryWithOpenCode });
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-42" });
+
+    const deleteLog = readFileSync(deleteLogPath, "utf-8");
+    expect(deleteLog).toContain("session delete ses_old_1");
+    expect(deleteLog).toContain("session delete ses_old_2");
+
+    expect(opencodeAgent.getLaunchCommand).toHaveBeenCalledWith(
+      expect.objectContaining({
+        projectConfig: expect.objectContaining({
+          agentConfig: expect.not.objectContaining({ opencodeSessionId: expect.any(String) }),
+        }),
+      }),
+    );
+
+    const metadata = readMetadataRaw(sessionsDir, session.id);
+    expect(metadata?.["opencodeSessionId"]).toBeUndefined();
+  });
+
+  it("throws for unknown project", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+    await expect(sm.spawn({ projectId: "nonexistent" })).rejects.toThrow("Unknown project");
+  });
+
+  it("throws when runtime plugin is missing", async () => {
+    const emptyRegistry: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockReturnValue(null),
+    };
+
+    const sm = createSessionManager({ config, registry: emptyRegistry });
+    await expect(sm.spawn({ projectId: "my-app" })).rejects.toThrow("not found");
+  });
+
+  describe("agent override", () => {
+    let mockCodexAgent: Agent;
+    let registryWithMultipleAgents: PluginRegistry;
+
+    beforeEach(() => {
+      mockCodexAgent = {
+        name: "codex",
+        processName: "codex",
+        getLaunchCommand: vi.fn().mockReturnValue("codex --start"),
+        getEnvironment: vi.fn().mockReturnValue({ CODEX_VAR: "1" }),
+        detectActivity: vi.fn().mockReturnValue("active"),
+        getActivityState: vi.fn().mockResolvedValue(null),
+        isProcessRunning: vi.fn().mockResolvedValue(true),
+        getSessionInfo: vi.fn().mockResolvedValue(null),
+      };
+
+      registryWithMultipleAgents = {
+        ...mockRegistry,
+        get: vi.fn().mockImplementation((slot: string, name: string) => {
+          if (slot === "runtime") return mockRuntime;
+          if (slot === "agent") {
+            if (name === "mock-agent") return mockAgent;
+            if (name === "codex") return mockCodexAgent;
+            return null;
+          }
+          if (slot === "workspace") return mockWorkspace;
+          return null;
+        }),
+      };
+    });
+
+    it("uses overridden agent when spawnConfig.agent is provided", async () => {
+      const sm = createSessionManager({ config, registry: registryWithMultipleAgents });
+
+      await sm.spawn({ projectId: "my-app", agent: "codex" });
+
+      expect(mockCodexAgent.getLaunchCommand).toHaveBeenCalled();
+      expect(mockAgent.getLaunchCommand).not.toHaveBeenCalled();
+    });
+
+    it("throws when agent override plugin is not found", async () => {
+      const sm = createSessionManager({ config, registry: registryWithMultipleAgents });
+
+      await expect(sm.spawn({ projectId: "my-app", agent: "nonexistent" })).rejects.toThrow(
+        "Agent plugin 'nonexistent' not found",
+      );
+    });
+
+    it("uses default agent when no override specified", async () => {
+      const sm = createSessionManager({ config, registry: registryWithMultipleAgents });
+
+      await sm.spawn({ projectId: "my-app" });
+
+      expect(mockAgent.getLaunchCommand).toHaveBeenCalled();
+      expect(mockCodexAgent.getLaunchCommand).not.toHaveBeenCalled();
+    });
+
+    it("persists agent name in metadata when override is used", async () => {
+      const sm = createSessionManager({ config, registry: registryWithMultipleAgents });
+
+      await sm.spawn({ projectId: "my-app", agent: "codex" });
+
+      const meta = readMetadataRaw(sessionsDir, "app-1");
+      expect(meta).not.toBeNull();
+      expect(meta!["agent"]).toBe("codex");
+    });
+
+    it("persists default agent name in metadata when no override", async () => {
+      const sm = createSessionManager({ config, registry: registryWithMultipleAgents });
+
+      await sm.spawn({ projectId: "my-app" });
+
+      const meta = readMetadataRaw(sessionsDir, "app-1");
+      expect(meta).not.toBeNull();
+      expect(meta!["agent"]).toBe("mock-agent");
+    });
+
+    it("uses project worker agent when configured and no spawn override is provided", async () => {
+      const configWithWorkerAgent: OrchestratorConfig = {
+        ...config,
+        projects: {
+          ...config.projects,
+          "my-app": {
+            ...config.projects["my-app"],
+            agent: "mock-agent",
+            worker: {
+              agent: "codex",
+            },
+          },
+        },
+      };
+
+      const sm = createSessionManager({
+        config: configWithWorkerAgent,
+        registry: registryWithMultipleAgents,
+      });
+      await sm.spawn({ projectId: "my-app" });
+
+      expect(mockCodexAgent.getLaunchCommand).toHaveBeenCalled();
+      expect(mockAgent.getLaunchCommand).not.toHaveBeenCalled();
+      expect(readMetadataRaw(sessionsDir, "app-1")?.["agent"]).toBe("codex");
+    });
+
+    it("uses defaults worker agent when project agent is not set", async () => {
+      const configWithDefaultWorkerAgent: OrchestratorConfig = {
+        ...config,
+        defaults: {
+          ...config.defaults,
+          worker: {
+            agent: "codex",
+          },
+        },
+        projects: {
+          ...config.projects,
+          "my-app": {
+            ...config.projects["my-app"],
+            agent: undefined,
+          },
+        },
+      };
+
+      const sm = createSessionManager({
+        config: configWithDefaultWorkerAgent,
+        registry: registryWithMultipleAgents,
+      });
+      await sm.spawn({ projectId: "my-app" });
+
+      expect(mockCodexAgent.getLaunchCommand).toHaveBeenCalled();
+      expect(readMetadataRaw(sessionsDir, "app-1")?.["agent"]).toBe("codex");
+    });
+
+    it("readMetadata returns agent field (typed SessionMetadata)", async () => {
+      const sm = createSessionManager({ config, registry: registryWithMultipleAgents });
+
+      await sm.spawn({ projectId: "my-app", agent: "codex" });
+
+      const meta = readMetadata(sessionsDir, "app-1");
+      expect(meta).not.toBeNull();
+      expect(meta!.agent).toBe("codex");
+    });
+  });
+
+  it("forwards configured subagent to spawn launch when no override is provided", async () => {
+    const configWithSubagent: OrchestratorConfig = {
+      ...config,
+      projects: {
+        ...config.projects,
+        "my-app": {
+          ...config.projects["my-app"],
+          agentConfig: {
+            subagent: "oracle",
+          },
+        },
+      },
+    };
+
+    const sm = createSessionManager({
+      config: configWithSubagent,
+      registry: mockRegistry,
+    });
+    await sm.spawn({ projectId: "my-app" });
+
+    expect(mockAgent.getLaunchCommand).toHaveBeenCalledWith(
+      expect.objectContaining({ subagent: "oracle" }),
+    );
+  });
+
+  it("prefers spawn subagent override over configured subagent", async () => {
+    const configWithSubagent: OrchestratorConfig = {
+      ...config,
+      projects: {
+        ...config.projects,
+        "my-app": {
+          ...config.projects["my-app"],
+          agentConfig: {
+            subagent: "oracle",
+          },
+        },
+      },
+    };
+
+    const sm = createSessionManager({
+      config: configWithSubagent,
+      registry: mockRegistry,
+    });
+    await sm.spawn({ projectId: "my-app", subagent: "librarian" });
+
+    expect(mockAgent.getLaunchCommand).toHaveBeenCalledWith(
+      expect.objectContaining({ subagent: "librarian" }),
+    );
+  });
+
+  it("validates issue exists when issueId provided", async () => {
+    const mockTracker: Tracker = {
+      name: "mock-tracker",
+      getIssue: vi.fn().mockResolvedValue({
+        id: "INT-100",
+        title: "Test issue",
+        description: "Test description",
+        url: "https://linear.app/test/issue/INT-100",
+        state: "open",
+        labels: [],
+      }),
+      isCompleted: vi.fn().mockResolvedValue(false),
+      issueUrl: vi.fn().mockReturnValue("https://linear.app/test/issue/INT-100"),
+      branchName: vi.fn().mockReturnValue("feat/INT-100"),
+      generatePrompt: vi.fn().mockResolvedValue("Work on INT-100"),
+    };
+
+    const registryWithTracker: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return mockAgent;
+        if (slot === "workspace") return mockWorkspace;
+        if (slot === "tracker") return mockTracker;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({
+      config,
+      registry: registryWithTracker,
+    });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-100" });
+
+    expect(mockTracker.getIssue).toHaveBeenCalledWith("INT-100", config.projects["my-app"]);
+    expect(session.issueId).toBe("INT-100");
+  });
+
+  it("succeeds with ad-hoc issue string when tracker returns IssueNotFoundError", async () => {
+    const mockTracker: Tracker = {
+      name: "mock-tracker",
+      getIssue: vi.fn().mockRejectedValue(new Error("Issue INT-9999 not found")),
+      isCompleted: vi.fn().mockResolvedValue(false),
+      issueUrl: vi.fn().mockReturnValue(""),
+      branchName: vi.fn().mockReturnValue("feat/INT-9999"),
+      generatePrompt: vi.fn().mockResolvedValue(""),
+    };
+
+    const registryWithTracker: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return mockAgent;
+        if (slot === "workspace") return mockWorkspace;
+        if (slot === "tracker") return mockTracker;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({
+      config,
+      registry: registryWithTracker,
+    });
+
+    // Ad-hoc issue string should succeed — IssueNotFoundError is gracefully ignored
+    const session = await sm.spawn({ projectId: "my-app", issueId: "INT-9999" });
+
+    expect(session.issueId).toBe("INT-9999");
+    expect(session.branch).toBe("feat/INT-9999");
+    // tracker.branchName and generatePrompt should NOT be called when issue wasn't resolved
+    expect(mockTracker.branchName).not.toHaveBeenCalled();
+    expect(mockTracker.generatePrompt).not.toHaveBeenCalled();
+    // Workspace and runtime should still be created
+    expect(mockWorkspace.create).toHaveBeenCalled();
+    expect(mockRuntime.create).toHaveBeenCalled();
+  });
+
+  it("succeeds with ad-hoc free-text when tracker returns 'invalid issue format'", async () => {
+    const mockTracker: Tracker = {
+      name: "mock-tracker",
+      getIssue: vi.fn().mockRejectedValue(new Error("invalid issue format: fix login bug")),
+      isCompleted: vi.fn().mockResolvedValue(false),
+      issueUrl: vi.fn().mockReturnValue(""),
+      branchName: vi.fn().mockReturnValue(""),
+      generatePrompt: vi.fn().mockResolvedValue(""),
+    };
+
+    const registryWithTracker: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return mockAgent;
+        if (slot === "workspace") return mockWorkspace;
+        if (slot === "tracker") return mockTracker;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({
+      config,
+      registry: registryWithTracker,
+    });
+
+    const session = await sm.spawn({ projectId: "my-app", issueId: "fix login bug" });
+
+    expect(session.issueId).toBe("fix login bug");
+    expect(session.branch).toBe("feat/fix-login-bug");
+    expect(mockTracker.branchName).not.toHaveBeenCalled();
+    expect(mockWorkspace.create).toHaveBeenCalled();
+  });
+
+  it("fails on tracker auth errors", async () => {
+    const mockTracker: Tracker = {
+      name: "mock-tracker",
+      getIssue: vi.fn().mockRejectedValue(new Error("Unauthorized")),
+      isCompleted: vi.fn().mockResolvedValue(false),
+      issueUrl: vi.fn().mockReturnValue(""),
+      branchName: vi.fn().mockReturnValue("feat/INT-100"),
+      generatePrompt: vi.fn().mockResolvedValue(""),
+    };
+
+    const registryWithTracker: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return mockAgent;
+        if (slot === "workspace") return mockWorkspace;
+        if (slot === "tracker") return mockTracker;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({
+      config,
+      registry: registryWithTracker,
+    });
+
+    await expect(sm.spawn({ projectId: "my-app", issueId: "INT-100" })).rejects.toThrow(
+      "Failed to fetch issue",
+    );
+
+    // Should not create workspace or runtime when auth fails
+    expect(mockWorkspace.create).not.toHaveBeenCalled();
+    expect(mockRuntime.create).not.toHaveBeenCalled();
+  });
+
+  it("spawns without issue tracking when no issueId provided", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+
+    const session = await sm.spawn({ projectId: "my-app" });
+
+    expect(session.issueId).toBeNull();
+    // Uses session/{sessionId} to avoid conflicts with default branch
+    expect(session.branch).toMatch(/^session\/app-\d+$/);
+    expect(session.branch).not.toBe("main");
+  });
+
+  it("sends prompt post-launch when agent.promptDelivery is 'post-launch'", async () => {
+    vi.useFakeTimers();
+    const postLaunchAgent = {
+      ...mockAgent,
+      promptDelivery: "post-launch" as const,
+    };
+    const registryWithPostLaunch: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return postLaunchAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({ config, registry: registryWithPostLaunch });
+    const spawnPromise = sm.spawn({ projectId: "my-app", prompt: "Fix the bug" });
+    await vi.advanceTimersByTimeAsync(5_000);
+    await spawnPromise;
+
+    // Prompt should be sent via runtime.sendMessage, not included in launch command
+    expect(mockRuntime.sendMessage).toHaveBeenCalledWith(
+      expect.objectContaining({ id: expect.any(String) }),
+      expect.stringContaining("Fix the bug"),
+    );
+    vi.useRealTimers();
+  });
+
+  it("does not send prompt post-launch when agent.promptDelivery is not set", async () => {
+    const sm = createSessionManager({ config, registry: mockRegistry });
+    await sm.spawn({ projectId: "my-app", prompt: "Fix the bug" });
+
+    // Default agent (no promptDelivery) should NOT trigger sendMessage for prompt
+    expect(mockRuntime.sendMessage).not.toHaveBeenCalled();
+  });
+
+  it("sends AO guidance post-launch even when no explicit prompt is provided", async () => {
+    vi.useFakeTimers();
+    const postLaunchAgent = {
+      ...mockAgent,
+      promptDelivery: "post-launch" as const,
+    };
+    const registryWithPostLaunch: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return postLaunchAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({ config, registry: registryWithPostLaunch });
+    const spawnPromise = sm.spawn({ projectId: "my-app" });
+    await vi.advanceTimersByTimeAsync(5_000);
+    await spawnPromise;
+
+    expect(mockRuntime.sendMessage).toHaveBeenCalledWith(
+      expect.objectContaining({ id: expect.any(String) }),
+      expect.stringContaining("ao session claim-pr"),
+    );
+    vi.useRealTimers();
+  });
+
+  it("does not destroy session when post-launch prompt delivery fails", async () => {
+    vi.useFakeTimers();
+    const failingRuntime: Runtime = {
+      ...mockRuntime,
+      sendMessage: vi.fn().mockRejectedValue(new Error("tmux send failed")),
+    };
+    const postLaunchAgent = {
+      ...mockAgent,
+      promptDelivery: "post-launch" as const,
+    };
+    const registryWithFailingSend: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return failingRuntime;
+        if (slot === "agent") return postLaunchAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({ config, registry: registryWithFailingSend });
+    const spawnPromise = sm.spawn({ projectId: "my-app", prompt: "Fix the bug" });
+    await vi.advanceTimersByTimeAsync(5_000);
+    const session = await spawnPromise;
+
+    // Session should still be returned successfully despite sendMessage failure
+    expect(session.id).toBe("app-1");
+    expect(session.status).toBe("spawning");
+    // Runtime should NOT have been destroyed
+    expect(failingRuntime.destroy).not.toHaveBeenCalled();
+    vi.useRealTimers();
+  });
+
+  it("waits before sending post-launch prompt", async () => {
+    vi.useFakeTimers();
+    const postLaunchAgent = {
+      ...mockAgent,
+      promptDelivery: "post-launch" as const,
+    };
+    const registryWithPostLaunch: PluginRegistry = {
+      ...mockRegistry,
+      get: vi.fn().mockImplementation((slot: string) => {
+        if (slot === "runtime") return mockRuntime;
+        if (slot === "agent") return postLaunchAgent;
+        if (slot === "workspace") return mockWorkspace;
+        return null;
+      }),
+    };
+
+    const sm = createSessionManager({ config, registry: registryWithPostLaunch });
+    const spawnPromise = sm.spawn({ projectId: "my-app", prompt: "Fix the bug" });
+
+    // Advance only 4s — not enough, message should not have been sent yet
+    await vi.advanceTimersByTimeAsync(4_000);
+    expect(mockRuntime.sendMessage).not.toHaveBeenCalled();
+
+    // Advance the remaining 1s — now it should fire
+    await vi.advanceTimersByTimeAsync(1_000);
+    await spawnPromise;
+    expect(mockRuntime.sendMessage).toHaveBeenCalled();
+    vi.useRealTimers();
+  });
+});
+
